refactor(teacher): TeacherHub sous 300 lignes — éclatement complet (H11)
Vue 418 → orchestrateur 68 l. Logique extraite dans useTeacherHub.js
(agrégation parallèle classes/matières/dashboard/séances). Template scindé
en HubNavCards (raccourcis) + HubQuickStats (aperçu rapide), CSS déplacé
verbatim avec leurs @media. Parité : mêmes appels services, mêmes classes
CSS, mêmes textes. Tests : composable (4) + 2 montages.
